# Release cuts in SVertexer