--- a/Detectors/Vertexing/include/DetectorsVertexing/SVertexerParams.h
+++ b/Detectors/Vertexing/include/DetectorsVertexing/SVertexerParams.h
@@ -47,13 +47,13 @@ struct SVertexerParams : public o2::conf::ConfigurableParamHelper<SVertexerParam
   bool usePropagator = false;                                           ///< use external propagator
   bool refitWithMatCorr = false;                                        ///< refit V0 applying material corrections
   //
-  int maxPVContributors = 2;              ///< max number PV contributors to allow in V0
-  float minDCAToPV = 0.1;                 ///< min DCA to PV of single track to accept
-  float minRToMeanVertex = 0.5;           ///< min radial distance of V0 from beam line (mean vertex)
-  float maxDCAXYToMeanVertex = 0.2;       ///< max DCA of V0 from beam line (mean vertex) for prompt V0 candidates
-  float maxDCAXYToMeanVertexV0Casc = 0.5; ///< max DCA of V0 from beam line (mean vertex) for cascade V0 candidates
-  float minPtV0 = 0.01;                   ///< v0 minimum pT
-  float maxTglV0 = 2.;                    ///< maximum tgLambda of V0
+  int maxPVContributors = 2;             ///< max number PV contributors to allow in V0
+  float minDCAToPV = 0.05;               ///< min DCA to PV of single track to accept
+  float minRToMeanVertex = 0.5;          ///< min radial distance of V0 from beam line (mean vertex)
+  float maxDCAXYToMeanVertex = 0.2;      ///< max DCA of V0 from beam line (mean vertex) for prompt V0 candidates
+  float maxDCAXYToMeanVertexV0Casc = 2.; ///< max DCA of V0 from beam line (mean vertex) for cascade V0 candidates
+  float minPtV0 = 0.01;                  ///< v0 minimum pT
+  float maxTglV0 = 2.;                   ///< maximum tgLambda of V0
 
   float causalityRTolerance = 1.; ///< V0 radius cannot exceed its contributors minR by more than this value
   float maxV0ToProngsRDiff = 50.; ///< V0 radius cannot be lower than this ammount wrt minR of contributors
@@ -68,18 +68,18 @@ struct SVertexerParams : public o2::conf::ConfigurableParamHelper<SVertexerParam
   float minRDiffV0Casc = 0.2; ///< cascade should be at least this radial distance below V0
   float maxRIniCasc = 90.;    // don't consider as a cascade seed (circles/line intersection) if its R exceeds this
 
-  float maxDCAXYCasc = 0.3; // max DCA of cascade to PV in XY // TODO RS: shall we use real chi2 to vertex?
-  float maxDCAZCasc = 0.3;  // max DCA of cascade to PV in Z
+  float maxDCAXYCasc = 0.5; // max DCA of cascade to PV in XY // TODO RS: shall we use real chi2 to vertex?
+  float maxDCAZCasc = 0.5;  // max DCA of cascade to PV in Z
   float minCosPACasc = 0.7; ///< min cos of PA to PV for cascade candidates
   float minPtCasc = 0.01;   // cascade minimum pT
   float maxTglCasc = 2.;    // maximum tgLambda of cascade
 
   // cuts on different V0 PID params
   bool checkV0Hypothesis = true;
-  float pidCutsPhoton[SVertexHypothesis::NPIDParams] = {0.001, 20, 0.60, 0.0};   // Photon
-  float pidCutsK0[SVertexHypothesis::NPIDParams] = {0.003, 20, 0.07, 0.5};       // K0
-  float pidCutsLambda[SVertexHypothesis::NPIDParams] = {0.001, 20, 0.07, 0.5};   // Lambda
-  float pidCutsHTriton[SVertexHypothesis::NPIDParams] = {0.0025, 14, 0.07, 0.5}; // HyperTriton
+  float pidCutsPhoton[SVertexHypothesis::NPIDParams] = {0.001, 20, 0.60, 0.0};    // Photon
+  float pidCutsK0[SVertexHypothesis::NPIDParams] = {0.003, 20, 0.07, 0.5};        // K0
+  float pidCutsLambda[SVertexHypothesis::NPIDParams] = {0.001, 20, 0.07, 0.5};    // Lambda
+  float pidCutsHTriton[SVertexHypothesis::NPIDParams] = {0.0025, 14, 0.07, 0.5};  // HyperTriton
   float pidCutsHhydrog4[SVertexHypothesis::NPIDParams] = {0.0025, 14, 0.07, 0.5}; // Hyperhydrog4 - Need to update
   //
   // cuts on different Cascade PID params
@@ -89,7 +89,6 @@ struct SVertexerParams : public o2::conf::ConfigurableParamHelper<SVertexerParam
 
   O2ParamDef(SVertexerParams, "svertexer");
 };
-
 } // namespace vertexing
 } // end namespace o2